# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "memorantado REST API",
    "version": "0.3.0",
    "description": "Local REST API used by the memorantado web UI."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3789"
    }
  ],
  "paths": {
...
    "/api/projects": {
      "get": {
        "summary": "List projects",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "summary": "Search graph and memory",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/graph": {
      "get": {
        "summary": "Read the project graph",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/entity": {
      "post": {
        "summary": "Create an entity",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/entity/{name}": {
      "get": {
        "summary": "Read an entity",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/entity/{name}/observations": {
      "post": {
        "summary": "Add an observation",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/observation/{id}": {
      "delete": {
        "summary": "Delete an observation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/relation": {
      "post": {
        "summary": "Create a relation",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/relation/{id}": {
      "delete": {
        "summary": "Delete a relation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/memory-items": {
      "get": {
        "summary": "List or search memory items",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a memory item",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/memory-items/{id}": {
      "get": {
        "summary": "Read a memory item",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a memory item",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes": {
      "get": {
        "summary": "List raw episodes",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Append an episode and optionally extract memories",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/episodes/{id}": {
      "get": {
        "summary": "Read an episode",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{id}/extract": {
      "post": {
        "summary": "Extract semantic memories",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/semantic-memories/{id}/lifecycle": {
      "post": {
        "summary": "Archive or restore a semantic memory",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/semantic-memories": {
      "get": {
        "summary": "List semantic memories",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or reinforce a semantic memory",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/semantic-memories/{id}/explain": {
      "get": {
        "summary": "Explain memory provenance",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/embeddings/backfill": {
      "post": {
        "summary": "Backfill configured embeddings",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "global",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "object",
                    "string",
                    "boolean",
                    "null"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "required": true
        }
      }
    },
...
  }
}